TRQ-3798. Make our decimal division for cgroup memory more precise.
Conflicts:
	src/test/numa_chip/scaffolding.c

+++ src/lib/Libutils/numa_chip.cpp
@@ -930,16 +930,16 @@ int Chip::free_core_count() const
  * @return the number of tasks that fit. This can be 0
  */
 
-float Chip::how_many_tasks_fit(
+double Chip::how_many_tasks_fit(
 
   const req &r,
   int        place_type) const
 
   {
-  float cpu_tasks;
-  float gpu_tasks;
-  float mic_tasks;
-  float mem_tasks = 0;
+  double cpu_tasks;
+  double gpu_tasks;
+  double mic_tasks;
+  double mem_tasks = 0;
 
   // Consider exclusive socket and node the same as exclusive chip for our purposes
   if ((place_type == exclusive_socket) ||
@@ -951,7 +951,7 @@ float Chip::how_many_tasks_fit(
        (this->chipIsAvailable()) == true))
     {
     // Need to handle place={core|thread}[=x]
-    float max_cpus = r.getExecutionSlots();
+    double max_cpus = r.getExecutionSlots();
     if (r.getPlaceCores() > 0)
       max_cpus = r.getPlaceCores();
     else if (r.getPlaceThreads() > 0)
@@ -964,7 +964,7 @@ float Chip::how_many_tasks_fit(
     else
       cpu_tasks = this->availableThreads / max_cpus;
 
-    long long memory = r.getMemory();
+    unsigned long memory = r.getMemory();
 
     // Memory isn't required for submission
     if (memory != 0)
@@ -978,7 +978,7 @@ float Chip::how_many_tasks_fit(
     else
       mem_tasks = cpu_tasks;
 
-    float gpus = r.getGpus();
+    double gpus = r.getGpus();
     if (gpus > 0)
       {
       gpu_tasks = this->available_gpus / gpus;
@@ -986,7 +986,7 @@ float Chip::how_many_tasks_fit(
         mem_tasks = gpu_tasks;
       }
 
-    float mics = r.getMics();
+    double mics = r.getMics();
     if (mics > 0)
       {
       mic_tasks = this->available_mics / mics;
